Allow for XEP-0184 support
XEP-0184 supports message delivery receipts.  Messages that request a
receipt send a <request /> element.  Messages that acknowledge receipt
send a message with a <received /> element and no <body> element.

This commit doesn't actually implement the XEP.  It just adds
necessary fields to the Message type so that an external component can
implement it itself.